Make encodeJSON and decodeJSON part of the codec class
