+++ elements/lisk-codec/src/codec.ts
@@ -18,6 +18,7 @@ import {
 	validator,
 	liskSchemaIdentifier,
 } from '@liskhq/lisk-validator';
+import { objects as objectUtils } from '@liskhq/lisk-utils';
 import { generateKey } from './utils';
 import { readObject, writeObject } from './collection';
 
@@ -30,6 +31,12 @@ import {
 	ValidatedSchema,
 	SchemaProps,
 } from './types';
+
+import {
+	iterator,
+	IteratableGenericObject,
+	recursiveTypeCast,
+} from './json_wrapper';
 
 export const validateSchema = (schema: {
 	// eslint-disable-next-line
@@ -112,6 +119,42 @@ export class Codec {
 
 		return (res as unknown) as T;
 	}
+
+	// For performance applications use decode() instead!
+	public decodeJSON = (schema: Schema, message: Buffer): GenericObject => {
+		const decodedMessage: IteratableGenericObject = this.decode(
+			schema,
+			message,
+		);
+		const decodedMessageCopy = objectUtils.cloneDeep(decodedMessage);
+		decodedMessageCopy[Symbol.iterator] = iterator;
+
+		recursiveTypeCast(
+			'toJSON',
+			decodedMessageCopy,
+			(schema as unknown) as SchemaProps,
+			[],
+		);
+		return decodedMessageCopy as GenericObject;
+	};
+
+	// For performance applications use encode() instead!
+	public encodeJSON = (
+		schema: Schema,
+		message: IteratableGenericObject,
+	): Buffer => {
+		const messageCopy = objectUtils.cloneDeep(message);
+		messageCopy[Symbol.iterator] = iterator;
+
+		recursiveTypeCast(
+			'fromJSON',
+			messageCopy,
+			(schema as unknown) as SchemaProps,
+			[],
+		);
+
+		return this.encode(schema, messageCopy);
+	};
 
 	private _compileSchema(
 		schema: ValidatedSchema | SchemaProps,
