+++ elements/lisk-codec/src/index.ts
@@ -14,4 +14,3 @@
 
 export * from './codec';
 export { Schema } from './types';
-export { encodeJSON, decodeJSON } from './json_wrapper';
